moved date calculations to DateUtils class
